+++ src/macros.rs
@@ -41,6 +41,50 @@ macro_rules! assert_csv_snapshot {
     }};
 }
 
+/// Asserts a `Serialize` snapshot in TOML format.
+///
+/// **Feature:** `toml` (disabled by default)
+///
+/// This works exactly like [`assert_toml_snapshot`](macro.assert_toml_snapshot.html)
+/// but serializes in [TOML](https://github.com/alexcrichton/toml-rs) format instead of
+/// YAML.  Note that TOML cannot represent all values due to limitations in the
+/// format.
+///
+/// Example:
+///
+/// ```no_run,ignore
+/// assert_toml_snapshot!(vec[1, 2, 3]);
+/// ```
+///
+/// The third argument to the macro can be an object expression for redaction.
+/// It's in the form `{ selector => replacement }`.  For more information
+/// about redactions see [redactions](index.html#redactions).
+///
+/// The snapshot name is optional but can be provided as first argument.
+/// For more information see [named snapshots](index.html#named-snapshots)
+#[cfg(feature = "toml")]
+#[macro_export]
+macro_rules! assert_toml_snapshot {
+    ($value:expr, @$snapshot:literal) => {{
+        $crate::_assert_serialized_snapshot!($value, Toml, @$snapshot);
+    }};
+    ($value:expr, {$($k:expr => $v:expr),*$(,)?}, @$snapshot:literal) => {{
+        $crate::_assert_serialized_snapshot!($value, {$($k => $v),*}, Toml, @$snapshot);
+    }};
+    ($value:expr, {$($k:expr => $v:expr),*$(,)?}) => {{
+        $crate::_assert_serialized_snapshot!($crate::_macro_support::AutoName, $value, {$($k => $v),*}, Toml);
+    }};
+    ($name:expr, $value:expr) => {{
+        $crate::_assert_serialized_snapshot!(Some($name), $value, Toml);
+    }};
+    ($name:expr, $value:expr, {$($k:expr => $v:expr),*$(,)?}) => {{
+        $crate::_assert_serialized_snapshot!(Some($name), $value, {$($k => $v),*}, Toml);
+    }};
+    ($value:expr) => {{
+        $crate::_assert_serialized_snapshot!($crate::_macro_support::AutoName, $value, Toml);
+    }};
+}
+
 /// Asserts a `Serialize` snapshot in YAML format.
 ///
 /// The value needs to implement the `serde::Serialize` trait and the snapshot