+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Added the `INSTA_SNAPSHOT_REFERENCES_FILE` environment variable to support
   deletions of unreferenced snapshot files. (#136)
+* Added support for TOML serializations.
 
 ## 1.0.0