+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 1.1.0
+
+* Added the `INSTA_SNAPSHOT_REFERENCES_FILE` environment variable to support
+  deletions of unreferenced snapshot files. (#136)
+* Added support for TOML serializations.
+
 ## 1.0.0
 
 * Globs now follow links (#132)